pyproject.toml
[project]
name = "dm"
version = "0.1.0"
description = "Dual momentum strategy calculator comparing VOO/VXUS returns against treasury rates"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fredapi>=0.5.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "yfinance>=1.0",
]

[project.scripts]
dm = "dm.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dm"]
